+++ tradinghub/backtest/trade_executor.py
@@ -1,16 +1,7 @@
-from dataclasses import dataclass
 from typing import Dict, Any, List
 import pandas as pd
 from tradinghub.models.dto.trade_results import Trade
-
-@dataclass
-class TradeParams:
-    """Parameters for trade execution"""
-    stop_loss_pct: float
-    take_profit_pct: float
-    max_holding_periods: int
-    commission: float
-    slippage: float
+from tradinghub.models.dto.trade_params import TradeParams
 
 class TradeExecutor:
     def __init__(self, trade_params: TradeParams):
